Milliseconds calculations overflowed on 32bit architectures. Thanks to wimschuermann for the findings and patch! It fixes bug #1132: http://bugs.cherokee-project.com/1132
git-svn-id: svn://cherokee-project.com/cherokee/trunk@6287 5dc97367-97f1-0310-9951-d761b3857238

+++ cherokee/bogotime.c
@@ -147,7 +147,7 @@ update_guts (void)
 
 	/* Update internal variables
 	 */
-	cherokee_bogonow_msec = ((cherokee_bogonow_tv.tv_sec * 1000) +
+	cherokee_bogonow_msec = (((unsigned long long)cherokee_bogonow_tv.tv_sec * 1000) +
 				 (cherokee_bogonow_tv.tv_usec) / 1000);
 
 	if (cherokee_bogonow_now == cherokee_bogonow_tv.tv_sec)
